porcupine/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
cycler==0.10.0
Cython==0.27.1
Django==1.11.6
django-custom-user==0.7
django-environ==0.4.1
django-filter==1.0.4
django-webpack-loader==0.4.1
-e git+https://github.com/mobolic/facebook-sdk.git@b953a62dba90dcce3dbea886fa90df93cf08cb70#egg=facebook_sdk
h5df==0.1.5
h5py==2.7.1
idna==2.6
iso8601==0.1.12
gunicorn==19.4.5
numpy==1.13.3
olefile==0.44
pandas==0.20.3
Pillow==4.3.0
promise==2.0.2
PyJWT==1.5.0
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
PyYAML==3.12
requests==2.18.4
singledispatch==3.4.0.3
six==1.11.0
typing==3.6.2
urllib3==1.22
waitress==1.0.1
whitenoise==3.2.2
cognitive-face==1.3.1
